.profile: removed if statement for ksh
For some reason, the .profile stopped exporting the appropriate
value for ENV, and so ksh wasn't picking up the kshrc file. Testing
the if statement itself manually confirmed that it did work... so
I don't know what the problem was. I'll need to investigate further
at some point. In the meantime, removing it works just fine.