Added Ubuntu's command not found functionality to shell
KSH now uses Ubuntu's command not found script to display the correct
install instructions for a command, if it isn't already installed when
you try to run it. The extraction of the command from history is a
little janky, though... I may revist that and try to make it a bit more
efficient.