sb-forecast: Temporary fix of undefined variable
This worked for a few days, and then randomly broke... Go figure.

This isn't the final form, but for now I've parameterized the
function that downloads the file, and set it up to use the
default location normally. That seems to work pretty well. It will
need further modification to allow the default to actually be
overridden, though.