screencasting: Added some scripts for screencasting
screencast toggles an FFMPEG screen recording on a specified
monitor. It should only allow a single recording to run at a
time using a status file in XDG_STATE_HOME.

This has also been integrated with the statusbar sb-indicators script
to include a little recording icon that is grey by default, and lights
up red when a recording is happening.

Finally, a convenience script to launch an MPV window viewing a webcam,
with low-latency options enable, is present as well.