get-pass: Added a flag for storing passwords unencrypted
I'm having trouble getting gpg-agent to work on OpenBSD from cron or
dwm, and so, at least for now, I've added the option to store passwords
in plaintext using the -c flag to get-pass. File format is the same as
the one I'm using for pass, but it doesn't require pinentry interaction
to get the password. Eventually, I hope to fix this, but all my OpenBSD
boxes are single-user anyway, so it isn't a catastrophic issue just yet.