nvim: revised completion keybindings
Tab was used to both navigate up and down the completion menu,
and also to advance to the next part of a snippet to fill in. This
effectively made snippets useless--as you'd need to quit out of the
completion dialog before you could advance to the next thing.

Now tab is strictly for advancing--the completion menu is navigated
with C-n and C-p. This'll take a bit of muscle memory retraining.